Dissecting ingest from `LogManager` (#746)
As in the description, moving ingest-related code to separate package,
reducing the responsibilities of `LogManager` object.

Because currently ingest depends on `SchemaRegistry`, it was impossible
(cyclic dependencies) to add `clickhouse.Table` type to LogManager.
Separating that will enable us to have all the table-related operations
done via the registry and gradually migrate to that direction.

Additionally, this refactor gave birth to `IngestProcessor`.