Run E2E tests as a required check in merge queue
Currently we don't run E2E tests on PRs or in the merge queue, just on
the main branch. That means that someone could merge a PR that
ultimately fails on the main branch. Fix that by running E2E tests
in the merge queue to catch the issue earlier.