Introduce auth  (#784)
This PR introduces an authentication for Elasticsearch API exposed by
Quesma, which - to put it short - defers to Elasticsearch when
authenticating incoming requests.
For more implementation details, there's an
[ADR](https://github.com/QuesmaOrg/quesma/blob/a46257bc70567016a4029837f28ba29ea32c03f0/adr/6_relying_on_elasticsearch_for_auth.MD).


### Notes for reviewer
1. I am adding a `disableAuth: true` option to the main Quesma
configuration so that our local dev setups maintain their lean nature 😉
2. I am also adding `elasticsearch.NewSimpleClient` as I have noticed
that creating a dumb `http.Client` which shoots ES API request occurs in
few places (`NewElasticJSONDatabase`, `NewIndexManagement`), so we can
reuse that logic. However, I'd prefer doing that refactor in a separate
PR, so for now only auth middleware is using it.

---------

Signed-off-by: Przemysław Hejman <przemyslaw.hejman@quesma.com>