Dual read configuration, default index configuration, disabling ingest for an index (#789)
This PR is dependent on #782. (first commit is from that PR)

This PR improves the index configuration in several ways:
- Dual read (A/B testing) can be enabled by specifying two targets in
query configuration of index:
  ```
  my_index:
    target: [clickhouse, elastic]
  ```
The first target is the "primary" target - the returned result from
Quesma will be from that target. Currently you can only specify
Clickhouse as the primary target (due to A/B testing limitations).
- Add mandatory default index configuration `*`. Before this
improvement, a user couldn't easily specify the behavior of Quesma for
non-configured indexes. Now the user can specify the configuration of
`*` index name, which will specify the default/fallback behavior. For
example, specifying this in the query configuration:
   ```
   "*":
      target: [clickhouse]
   ```
   will enable autodiscovery.
- Make it possible to disable ingest for an index by specifying empty
target list (`target: []`)

---------

Signed-off-by: Przemysław Hejman <przemyslaw.hejman@quesma.com>
Co-authored-by: Przemysław Hejman <przemyslaw.hejman@quesma.com>