Revert `*` functionality (#817)
Revert the `*` functionality in index configuration. For the time being
(for the next release), users of Quesma will have to explicitly mention
all enabled indexes and autodiscovery is not supported.

Please note that as of this commit, if some index X is not configured,
Quesma will route queries/ingest to index X into ElasticSearch.