Change default log level to INFO
Before this change if someone did not configure the log level in the
configuration it would default to DEBUG (as this is the first entry
in DebugLevel). Debug log level is very verbose and it's not a great
experience for a user (who tried a minimal config).

Set log level to INFO if none was provided in the loaded config.