Remove requirement of defining indexes in both processors, bring back required `*` configuration (#822)
Remove the requirement that the user has to configure some index in both
ingest and query processors. After this change, the user can now define
an index only in ingest or only in query processor.

Under the hood, the translated V1 index configuration will have an
ingest target set to `DefaultIngestTarget` if index is missing from
ingest processor configuration or query target set to
`DefaultQueryTarget` if index is missing from query processor
configuration.

Bring back required `*` configuration. The only supported configuration
is `"*": target: [elastic]`. This will make it easier for the users to
understand the configuration, instead of relying on some "magic"
fallback behavior.

Additionally improve the logging of index configurations (printed at the
start of Quesma).