Unskip 2 tests + fix to `min_doc_count=0` in `date_histogram` (#838)
We had 2 tests skipped, they weren't trivial to fix in pre-pancake era,
but with pancakes they should work out of the box.

One of them didn't, because for `date_histogram` we kind of have 2
different keys. One original that we get from DB, and one after
processing, which is a proper response to Kibana.
During merging of a `date_histogram` with its subaggregation, we need
the original one, as it's the only one that subaggregation has, and we
do merging comparing those 2 keys.
It's not easy to get the original one from the processed one, so I just
cache the original one in the response, similar to `__Quesma_key` I
introduced long time ago (which I'll try to remove now in another PR)