Fix problems with dates part 1 (#832)
Query from the test caused Quesma to crash without any error message, as
it'd run of out memory in 0,1s or so. You can see it below:
![Screenshot 2024-10-02 at 15 33
54](https://github.com/user-attachments/assets/c1a4cda2-a302-4127-b6d4-570784060e4e)
![Screenshot 2024-10-02 at 15 34
09](https://github.com/user-attachments/assets/4e5565b8-829d-43fe-9ad7-3d1a4cf30faa)
It was a small bug when adding empty rows. Usually keys in
`date_histogram` are `x, x+1, x+2,...`, when we're using `timestamp /
duration` as a key, so number of added empty rows was kind of limited.
But in `calendar_interval` for big intervals, like 1 week, keys are
timestamps in millisecond, so we were adding a new row for each
millisecond, exhausting all the memory very quickly.
Fixed here by considering the second option.