small improvement `top_hits` (#848)
* Fixes some small problem with not having count in `top_hits` pancake.
We can't get it (at least easily) in our logic, as you can't do `SELECT
field, count()`. But it doesn't seem to be a problem, it's of course
present in some other pancake and it's only taken from there => seems
working fine.
* Add some better support for what fields we want to return in
`top_hits`. Before it had to be an array `params[_source][includes]`,
now it can also. be a string `params[_source]`. It's not 100% support
for this `_source` field, but a bit better.

---------

Co-authored-by: Jacek Migdal <jacek@quesma.com>