Extract `build-quesma-docker-image` to a reusable GitHub Actions workflow (#862)
While working on an upcoming PR, I noticed that
`build-quesma-docker-image` step in GitHub Actions workflow was
duplicated in 3 places... and my PR would add a 4th copy!

Therefore extract the `build-quesma-docker-image` step into a "reusable
workflow" (`workflow_call` trigger) which can be called from other
workflows, avoiding duplication. All the behaviors of this step stay the
same - for now the "build+push" step and "build+export" step are
separate to make sure the behavior stays the same.

Additionally `docker-image-latest.yml` is renamed to
`nightly-docker-build-and-push.yml` in order to avoid confusion with the
new `build-quesma-docker-image.yml` file.

---------

Signed-off-by: Piotr Grabowski <avelanarius@gmail.com>