Run integration tests against `nightly` (not `latest`) (#865)
The `latest` tag of `quesma/quesma` is the latest released stable
release of Quesma, while `nightly` is the newest build of Quesma.
Integration tests used to run against `quesma/quesma:latest` and not the
last `nightly` build of Quesma, which meant that we wouldn't catch some
failures before release.

Fix this issue by running the integration tests against `nightly` and
always building the newest `quesma/quesma:nightly` in the
`integration-test.yml` GitHub workflow.