Fixing quesma common table dynamic mapping case, second approach (#885)
This is a second approach (1 -
https://github.com/QuesmaOrg/quesma/pull/878) of fixing a problem where
mappings contains more fields than ingest and sql queries are based on
this knowledge. In this approach we are not altering table to have all
fields that were added by dynamic mapping but we based on field origin
information and building queries according to that.

<img width="1706" alt="image"
src="https://github.com/user-attachments/assets/40e52e56-7da5-4346-b138-8f5b738513c4">