Allow more configurations of targets for "*" (#905)
Previously, a user was allowed to only put ElasticSearch connector as a
target for `*` index configuration (describing the default/fallback
behavior). This PR lifts that requirement and allows for any connectors
to be put into ingest's `*` and at most any 1 connector into query's
`*`.

`table_resolver` code is modified to now read the `*` configuration and
use it for decision (instead of always falling back to Elastic).

This PR also adds back enabling autodiscovery via configuration - if
`ClickHouse` target is listed in `*` targets, autodiscovery will be
turned on.