Add test for A/B testing table_resolver bug (in case of pattern)
A/B testing is turned on in the configuration by specifying
two targets: ClickHouse and Elastic (`target: [ch, es]`). table_resolver
correctly handles it for single index case, but it's not working
correctly for the pattern case (even if the pattern matches only
a single index with A/B testing requested).

Adding a (skipped) test for such case (a proper fix needs
some discussion).