Lucene enhancements (#921)
This Lucene parser seemed to work fine for a subset of Lucene. Now I had
to extend this subset a bit, and I'm afraid some weird (and usually
incorrect) queries might not work 100% correctly, although I'd have to
think for a bit for an example.

But all tests + customer's dashboards work fine, so I'd merge it. I have
an idea how to support full Lucene, 90% it'd work. I'll try to do it
very soon.

I tried 2 other open-source Lucene parsers, and they both fail on ~20-30
/ 60 our tests, so if our passes all, it's more fine than not.