Make ClickHouse connection configuration more user-friendly
Before this change, Quesma supported only the "clickhouse://" protocol
to connect to ClickHouse. However, ClickHouse also supports the HTTP
protocol (8123) and HTTPS protocol (8443). This PR extends the
ClickHouse connection logic to switch to using HTTP(S) protocol if the
user put that endpoint instead.

This change is motivated by the fact that the HTTPS endpoint is the
first/default endpoint shown on ClickHouse Cloud and by the fact that
we saw a user try to use the HTTP endpoint.