Add "ClickHouse engine" limitation (#936)
Currently Quesma auto-creates tables with `MergeTree` engine. It's
automatically converted to `ReplicatedMergeTree` in case of ClickHouse
Cloud, but not in case of a local multi-node cluster and the user can't
control it.

If the user has a local multi-node cluster, the tables automatically
created by Quesma won't be replicated - the only solution is to create
them manually with the correct `ReplicatedMergeTree` engine.

Refs #934

Signed-off-by: Piotr Grabowski <piotr@quesma.com>