Replace all `target` configurations with new variant (#1015)
In #945, a new ("extended") target configuration was introduced:

```yaml
target:
  - a
  - b
```

instead of `target: [a, b]`.

We still maintain compatibility with the old target configuration.
However, we should educate users (in our docs) to use the new target
configuration.

This PR replaces all, internal and external (docs), uses of old target
configuration with the new ("extended") target configuration. A new test
(`TestTargetLegacyVariant`) is added to make sure we still maintain
compatibility with the old way of declaring targets.