Wait for HTTP 200 OK in integration tests
We sometimes see integration tests failing, because Quesma returns
HTTP 503 Service Unavailable and the test tries to use Quesma in such
a unhealthy state. (The only place that 503 could originate from is
the golang's http library code - in a case that a handler timeouts.
I didn't think it was worthwhile to investigate it further, we only
saw this issue in integration tests, not ever in a normal operation of
Quesma).

This PR fixes the WaitingFor condition to wait for HTTP 200 OK for
Elastic and Quesma containers. This should prevent the test from
starting when the container is not yet ready.