Fix NOTICE.MD generator for quesma_v2 "dependency"
We are in the process of an internal code refactor and we have
introduced a new module "quesma_v2" for this refactor.

Unfortunately, the go-licence-detector (generating NOTICE.MD) tries
to find a licence for quesma_v2 module and it fails.

We could inform it that "quesma_v2" is licensed under "Elastic-2.0",
however that licence is not allowed in the default go-licence-detector
configuration. I wasn't sure if we should allow "Elastic-2.0"
dependencies generally, therefore this change works around this issue
by just filtering out "quesma_v2" from dependency list.