Print container logs when integration test fails
We see (relatively rarely) integration tests being flaky. In those
cases of flakiness the current logging is not sufficient to really
understand the cause of the problem.

This PR aims to improve logging by printing logs of all containers when
a test fails.

printContainerLogs is added to Containers.Cleanup() code to print the
logs of containers.

Additionally, the Cleanup() function is now ALWAYS called, even if
the containers failed to start or tests failed. To make sure it can
print some logs:
- the code now has less panics, which would have prevented Cleanup()
from running
- in setupClickHouse() and others we return the container object in case
container failed to start instead of nil (but still with error) to be
able to recover some logs
- if one container fails to start, Containers is now returned partially
filled (but still with error), so that we can print some logs instead of
giving up entirely