Async Insert Optimizer (#1151)
This PR adds the ability to buffer inserts. If the optimizer is enabled,
we add an "async insert"[1] setting to every `INSERT` query.

Motivation: clickhouse doesn't handle multiple inserts per second well.
We need to compensate for this.


The optimizer can be enabled for all indexes 
```
...
# ingest pipeline config 

...
        '*':
          optimizers:
            - async_insert:
                disabled: false
                properties:
                  wait_for_async_insert: 0 
                  async_insert_busy_timeout_ms: 10000
          target: [ my-clickhouse-data-source ]
...
```
Or can be added to a single index.
```
...
        device_logs:
          target: [ my-clickhouse-data-source ]
          optimizers:
            - async_insert:
                disabled: true

        other:
          target: [ my-clickhouse-data-source ]
          optimizers:
            - async_insert:
                disabled: false
...
```




[1] https://clickhouse.com/blog/asynchronous-data-inserts-in-clickhouse