`date_histogram`: fix adding empty in-between rows (#1158)
We had a problem when adding empty rows for `>= month`
`calendar_interval`, as we always added fixed amount of time, which
obviously isn't right for `month/quarter/year` interval.
This PR fixes that but only for `in-between` rows. Tried also `pre` and
`post`, but that wasn't as trivial, because our `date_histogram` is the
most complex aggregation, I think, and its code is a bit of a mess. I'll
finish that in next PR.