Fix misrouting of /_license (and others) to ClickHouse
In ElasticSearch there's a "GET /:index" endpoint. Quesma supports that
endpoint, however an "GET /_license" request erroneously matched to
that endpoint, routing it into ClickHouse instead of a fallback
ElasticSearch.

Previously a couple endpoints (such as "/_nodes", "/_xpack", "/_stats")
were hardcoded to be routed to Elastic, however that list was not
exhaustive and was prone to become obsolete if Elastic introduced a new
endpoint.

This PR changes the routing logic - now if an index name starts with
"_" (forbidden) it is assumed that it is an internal path and should
be routed to Elastic.