Fix misrouting of `/_license` (and others) to ClickHouse (#1225)
In ElasticSearch there's a `GET /:index` endpoint. Quesma supports that
endpoint, however a `GET /_license` request erroneously matched to that
endpoint, routing it into ClickHouse instead of a fallback to
ElasticSearch.

Previously a couple similar endpoints (such as `/_nodes`, `/_xpack`,
`/_stats`) were hardcoded to be routed to Elastic, however that list was
not exhaustive and was prone to become obsolete if Elastic introduced a
new endpoint.

This PR changes the routing logic - now if an index name starts with `_`
(forbidden index name in Elastic and indicative of an endpoint name) it
is assumed that it is an internal path and should be routed to Elastic.