Improve performance, reliability of MySQL (TCP) connection handling
Firstly, this PR improves the performance of the MySQL TCP connection
handling. When a MySQL client sends a request, we forward it to a real
MySQL server and then in loop try to read all response packets.
Previously we didn't have a good mechanism to detect the final response
packet and relied on a read timing out (1s), which was slow for
Wordpress which sends 10s or 100s of requests to render a page.

Secondly, this PR improves the reliability of connection handling:
we no longer leak opened connections, we now have better logging
and I have confirmed manually that it's really stable on a  Wordpress
instance.