Initial version of MySQL frontend connector (based on Vitess) (#1212)
In #1211, an initial (experimental) version of (TCP) passthrough MySQL
processor was introduced. That processor/frontend was merely a minimally
invasive TCP proxy, which snooped on the traffic, was able to recognize
Query packets and potentially modify some packets. That
processor/frontend required us to always have a real MySQL instance
running and it was not easy for the processor to use a different SQL DB
as a backend (such as Postgres).

The components in this PR are MySQL frontend connector (based on Vitess)
and MySQL processor (paired with MySQL Vitess frontend connector). They
try to emulate a MySQL instance and don't require having a real MySQL
instance running. A disadvantage of this approach is that we have to
manually construct MySQL responses (abstracted by Vitess structures).

This is still an experimental frontend/processor (alpha) and it's only a
base code, on which we will further experiment.