Add Unlicence to allowed licence list for go-licence-detector
After #1212, go-licence-detector GitHub Actions job started failing.
One of the added dependencies uses Unlicence licence, which is not
on the allowlist of the default go-licence-detector rules list.

Unlicence is a public-domain-equivalent licence. The default rules list
already allowed similar licences, such as "CC0-1.0" and "Public Domain",
so it should be OK to add Unlicence to the allowed list.

The added rules.json is a default rules file with only Unlicence added.

This fixes the NOTICE.MD generation with go-licence-detector.