Group indexes by prefix (common table) (#1175)
When Grafana queries daily indexes:
- it adds a time range 
- it adds a list of the indexes within a time range 

So if we have a 3-year time range, Grafana will send 1k+ indexes in the
query. We transform that list into an `OR` expression, and hit the query
limit.

This PR adds an optimizer that transforms the `OR` expressions into a
single `starsWith` expression when indexes share the same prefix.

If we have indexes with a pattern `daily-YYYYMMDD,` we can enable the
optimizer:
```
        '*':
          useCommonTable: true
          target: [ my-clickhouse-data-source ]
          optimizers:
              - group_common_table_indexes:
                  disabled: false
                  properties:
                        "daily-": true
```

---------

Signed-off-by: Rafał Strzaliński <rafal@quesma.com>