Improve `_msearch` handling (#1251)
This PR puts a tiny improvement over multi-search handling: 

❗ `:index/_msearch` calls targeted to Elasticsearch will be routed
directly to Elasticsearch instead of failing later on with enigmatic `no
clickhouse connector` error.

Implementing a high-road fix in the current architecture requires adding
a somehow 'dirty' dependency (QueryRunner will have to call
Elasticsearch directly). Overall we can do it and sanitize
`handleSearchCommon` a little too, but together with @pdelewski we
decided to postpone this for now.