Skip invalid/incomplete types in CREATE TABLE
When a user sends a JSON with a nil field, we can't determine a
ClickHouse type for such field (column). This is the case for nil
(we used to fall back to String) an empty arrays or empty dictionaries.

This PR changes the logic in NewType() - when the type is determined
to be invalid/incomplete, the function now returns an error. The
CREATE TABLE code now can skip such problematic column and hopefully
it will be handled in some future ALTER TABLE ADD COLUMN statement
(once Quesma sees an example with complete type).