Add Unlicence to allowed license list for go-licence-detector (#1236)
After #1212, go-licence-detector GitHub Actions job started failing. One
of the added dependencies uses
[Unlicense](https://en.wikipedia.org/wiki/Unlicense) license, which is
not on the allowlist of the default go-licence-detector rules list.

Unlicense is a public-domain-equivalent license. The default rules list
already allowed similar licenses, such as "CC0-1.0" and "Public Domain",
so it should be OK for us to add Unlicense to the allowed list.

The added rules.json is a default rules file with only Unlicense added.

This fixes the NOTICE.MD generation with go-licence-detector.