Rate limit logs globally (#1262)
This is a first PR in a series to reduce amount of logs generated by
Quesma. This PR rate limits logs globally for entire Quesma. This is
meant as a fail-safe in case some codepath generates too much logs
(specific codepaths will be fixed in next PRs).

The rate limits in this PR are generous, as it's only a fail-safe:
- ~100k log lines per hour
- Allow for bursts of 600 lines of logs (per 20 second period)
- When the limit is exhausted (for the 20 second period), logger will
cut down on amount of logs, printing 1/100th of logs. This shedding is
not an elegant solution, but that way the user sees that Quesma is still
running and printing something (until the next 20 second period).