Add deduplicated log line printing to quesma_logger, use it in places that logged most often (#1266)
This PR introduces `logger.DeduplicatedWarn()` and
`logger.DeduplicatedInfo()` methods to `quesma_logger`, which
automatically deduplicate log messages.

For example,

```
  Ingesting via _bulk API, batch size=2 lines
```

is a log line that the code tries to print very often, but with
`DeduplicatedInfo()` that exact log line is only printed once. An
expirable LRU cache is used with 1 minute expiry time, so the log line
is still printed occasionally, but at most 1 every minute.

Additionally, a couple places that produced a lot of logs were migrated
to use `DeduplicatedWarn()`/`DeduplicatedInfo()`.

---------

Signed-off-by: Piotr Grabowski <piotr@quesma.com>